pyproject.toml
[project]
name = "Bagels"
version = "0.1.3"
authors = [
    { name = "Jax", email = "enhancedjax@gmail.com" }
]
description = "Powerful terminal expense tracker."
readme = "README.md"
requires-python = ">=3.13"
license = { file = "LICENSE" }
keywords = ["python", "cli", "expense-tracker", "python3", "UV", "ssh", "textual", "terminal", "tui"]
dependencies = [
    "aiohappyeyeballs==2.4.3",
    "aiohttp-jinja2==1.6",
    "aiohttp==3.10.10",
    "aiosignal==1.3.1",
    "annotated-types==0.7.0",
    "attrs==24.2.0",
    "blinker==1.8.2",
    "click-default-group>=1.2.4",
    "click==8.1.7",
    "flask-sqlalchemy==3.1.1",
    "flask==3.0.3",
    "freezegun>=1.5.1",
    "frozenlist==1.5.0",
    "idna==3.10",
    "itsdangerous==2.2.0",
    "jinja2==3.1.4",
    "linkify-it-py==2.0.3",
    "markdown-it-py==3.0.0",
    "markupsafe==3.0.2",
    "mdit-py-plugins==0.4.2",
    "mdurl==0.1.2",
    "msgpack==1.1.0",
    "multidict==6.1.0",
    "platformdirs==4.3.6",
    "plotext==5.3.2",
    "propcache==0.2.0",
    "pydantic-core==2.23.4",
    "pydantic-settings-yaml>=0.2.0",
    "pydantic-settings>=2.6.1",
    "pydantic==2.9.2",
    "pygments==2.18.0",
    "pytest>=8.3.3",
    "pyyaml==6.0.2",
    "rich==13.9.3",
    "sqlalchemy==2.0.36",
    "textual-dev==1.6.1",
    "textual==0.84.0",
    "tomli>=2.0.2",
    "typing-extensions==4.12.2",
    "uc-micro-py==1.0.3",
    "werkzeug==3.0.6",
    "xdg-base-dirs>=6.0.2",
    "yarl==1.16.0",
]

[project.urls]
Homepage = "https://github.com/EnhancedJax/Bagels"
Repository = "https://github.com/EnhancedJax/Bagels"
Issues = "https://github.com/EnhancedJax/Bagels/issues"
Documentation = "https://github.com/EnhancedJax/Bagels/blob/main/README.md"


[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project.scripts]
bagels = "bagels.__main__:cli"

[tool.uv]
dev-dependencies = [
    "textual-dev>=1.5.1",
    "pytest>=8.3.1",
    "jinja2>=3.1.4",
    "syrupy>=4.6.1",
    "pytest-xdist>=3.6.1",
    "pytest-cov>=5.0.0",
    "pytest-textual-snapshot>=1.0.0",
    "mkdocs-material>=9.5.30",
]

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/bagels"]

[tool.pytest.ini_options]
markers = [
    "serial",  # used to indicate tests must not run in parallel
]

[tool.coverage.run]
relative_files = true

